Extract spherical geography functions to SphericalGeoFunctions
1. This reduces the size of GeoFunctions, and keeps spherical functions
in one place.
2. SqlTypes Geometry and SphericalGeography have the same Java type
(Slice).  This makes it impossible to put functions that overload
between Geometry and SphericalGeography in one class.  With
SphericalGeoFunctions, we can overload the SqlTypes in two different
classes.
