+++ presto-geospatial/src/main/java/com/facebook/presto/plugin/geospatial/GeoPlugin.java
@@ -50,6 +50,7 @@ public class GeoPlugin
                 .add(KdbTreeCasts.class)
                 .add(SpatialPartitioningAggregateFunction.class)
                 .add(SpatialPartitioningInternalAggregateFunction.class)
+                .add(SphericalGeoFunctions.class)
                 .build();
     }
 }
